+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.4] 2025-03-17
+
+### Changed
+- Add default object names for new objects (object type + object, e.g. UsageJourney 3).
+
 ## [0.6.3] 2025-03-17
 
 ### Fixed